+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 - No longer required `agent` from the `/auth` request and the client itself.
 - No longer supported `/mod/meta` request.
+- No longer supported `isEnabled` flag from the `Mod` entity.
 
 ## 2.1.0 - 2019-07-15